Add log to frontend and change downloadVideo to be a promise
To add logging to the front end, we need to add an event listener
to the front end. There are two levels of errors:

a) Error on the front end (e.g. no link provided or dialog cancelled)
b) Error from the Python api

Errors for (a) are handled in main.js. Errors for (b) are handled using
a response handler handleResponse.js. downloadVideo.js needs to be modified
to return a promise which result is the status of the api.